--- springboot-lab/DevTeam2/bagay/api-demo/.mvn/wrapper/maven-wrapper.properties
@@ -1,3 +1,0 @@
-wrapperVersion=3.3.4
-distributionType=only-script
-distributionUrl=https://repo.maven.apache.org/maven2/org/apache/maven/apache-maven/3.9.14/apache-maven-3.9.14-bin.zip
