--- springboot-lab/DevTeam2/bagay/api-demo/ci-stubs/org/junit/jupiter/api/Test.java
@@ -1,5 +1,0 @@
-package org.junit.jupiter.api;
-
-public @interface Test {
-}
-
