--- springboot-lab/DevTeam2/bagay/api-demo/ci-stubs/org/springframework/boot/SpringApplication.java
@@ -1,11 +1,0 @@
-package org.springframework.boot;
-
-public final class SpringApplication {
-    private SpringApplication() {
-    }
-
-    public static void run(Class<?> primarySource, String... args) {
-        // CI stub for repository-wide javac check.
-    }
-}
-
