--- springboot-lab/DevTeam2/bagay/api-demo/ci-stubs/org/springframework/boot/autoconfigure/SpringBootApplication.java
@@ -1,5 +1,0 @@
-package org.springframework.boot.autoconfigure;
-
-public @interface SpringBootApplication {
-}
-
