--- springboot-lab/DevTeam2/bagay/api-demo/ci-stubs/org/springframework/web/bind/annotation/GetMapping.java
@@ -1,6 +1,0 @@
-package org.springframework.web.bind.annotation;
-
-public @interface GetMapping {
-    String[] value() default {};
-}
-
