--- springboot-lab/DevTeam2/bagay/api-demo/ci-stubs/org/springframework/web/bind/annotation/RestController.java
@@ -1,5 +1,0 @@
-package org.springframework.web.bind.annotation;
-
-public @interface RestController {
-}
-
