--- springboot-lab/DevTeam2/bagay/api-demo/src/main/java/com/bagay/springboot/ApiDemoApplication.java
@@ -1,13 +1,0 @@
-package com.bagay.springboot;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ApiDemoApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ApiDemoApplication.class, args);
-	}
-
-}
