--- springboot-lab/DevTeam2/bagay/api-demo/src/main/resources/application.properties
@@ -1,11 +1,0 @@
-spring.application.name=Bagay Spring Boot API
-
-# Server configuration
-server.port=8080
-
-# Application info
-app.name=Bagay Spring Boot API
-app.version=1.0.0
-
-# Logging
-logging.level.com.bagay.springboot=DEBUG
