--- springboot-lab/DevTeam2/bagay/api-demo/src/test/java/com/bagay/springboot/ApiDemoApplicationTests.java
@@ -1,13 +1,0 @@
-package com.bagay.springboot;
-
-import org.junit.jupiter.api.Test;
-class ApiDemoApplicationTests {
-
-	@Test
-	void contextLoads() {
-		if (ApiDemoApplication.class == null) {
-			throw new AssertionError("ApiDemoApplication class is missing");
-		}
-	}
-
-}
